SYSTEST-9456-addressed comment to move functions to util file
